searchall command
This is similar to the "apt-file search" command, and almost identical to the
"cygcheck --package-query" command. Instead of only search for a file from
installed packages, it searches all available packages.

I have included it as a separate command because it "wget"s cygwin.com and can
be slow at times.