apt-msys2
This is a file similar to apt-cyg. It will install packages from the MSYS2
project. It will only be practical for packages from the "MINGW" repo, as
packages from the "MSYS2" repo would require the MSYS2 dll.

I have a need for this because Cygwin Tcl/Tk requires X11. MSYS2 Tcl/Tk does not
require X11, so it can run much faster.