Remove wget dependency
apt-cyg has had a long standing circular dependency with wget. This is because
Base Cygwin does not include wget or curl. I have mentioned before workarounds
to this

41e1d911728653eaf314769ea29ea94df69a3702

However none were ideal. I have since discovered lynx, which can be used in a
pinch to download files

http://unix.stackexchange.com/a/83987

lynx has these features

- pure command line, unlike setup-x86_64
- works with "https://" URLs, unlike /dev/tcp
- comes with Base Cygwin, unlike wget/curl

If wget is not installed, lynx will be used as a fallback with a warning
printed.