Begin Go rewrite
This may end up being nothing, but would like to rewrite apt-cyg in Go. This
will be nice because you will end up with a single EXE file, similar to
setup-x86_64.exe. The difference is that it will be totally command line like
apt-cyg, but better because you can build all the requirements
(wget, bunzip2, awk) right into the EXE. My concern right now is that I will
probably have to write an INI parser, because Cygwin setup.ini is a weird
format. Since I am just starting with Go I may not have the skill, we shall see.